+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 #### 1.1.0 - 2014-10-29
  - Renamed project to Stork.
  - Improved docs
+ - Support for directories and wildcards with input file arguments
 
 #### 1.0.1 - 2014-10-29
  - Reorganized project layout in sub-modules.